hddtemp: improve detection of drives
Previously, this script was only able to find 26 drives (sda-sdz) due to the use of globbing.

A better strategy for detecting drives would be to use lsblk on systems that support it, failing over to globbing.

This patch adds support both for lsblk and a more comprehensive glob solution with find that will at least catch 26^2 drives.