ntp-client data correction (#196)
NTP was not displaying data right for linux servers. It was putting the frequency data into the offset data. This was giving bad graphs in the UI. Tested the correction on both RHEL and Debian based operating systems and all passes.

Remove the .sh to simplify for configuration management orchestration scripts.